Provide list of public consts
libe2p does not provide a stable API. The set of defined constants
depends on the exact version. Crates using these bindings need a way
to figure out which constants exist. We provide such a way by collecting
their names in an array of &str.

Minor change: Use unwrap() instead of expect() in build.rs and in the
tests. At build and test time we can be a bit lazy here.